feat(tms): 整合头程申请单与船期信息- 在 TmsFirstMileRequestSaveReqVO 中添加船期信息字段 - 更新 TmsFirstMileRequestRespVO，增加船期信息 - 修改 TmsFirstMileController 和 TmsFirstMileItemMapper 以支持船期查询 - 更新相关 DTO 和 VO 类以适应新的船期信息字段

+++ yudao-module-tms/yudao-module-tms-biz/src/main/java/cn/iocoder/yudao/module/tms/controller/admin/first/mile/vo/req/TmsFirstMilePageReqVO.java
@@ -162,7 +162,7 @@ public class TmsFirstMilePageReqVO extends PageParam {
         private String voyage;
 
         @Schema(description = "货代公司(供应商ID)")
-        private Integer forwarderCompanyId;
+        private Long forwarderCompanyId;
 
         @Schema(description = "箱号")
         private String containerNo;
