+++ yudao-module-tms/yudao-module-tms-biz/src/main/java/cn/iocoder/yudao/module/tms/controller/admin/vessel/tracking/vo/TmsVesselTrackingPageReqVO.java
@@ -83,7 +83,7 @@ public class TmsVesselTrackingPageReqVO extends PageParam {
     private String voyage;
 
     @Schema(description = "货代公司(供应商ID)")
-    private Integer forwarderCompanyId;
+    private Long forwarderCompanyId;
 
     @Schema(description = "箱号")
     private String containerNo;
